Refactor Equipment class.
