--- game/src/main/kotlin/xlitekt/game/actor/player/Equipment.kt
@@ -1,17 +1,0 @@
-package xlitekt.game.actor.player
-
-class Equipment {
-    companion object {
-        const val SLOT_HEAD = 0
-        const val SLOT_CAPE = 1
-        const val SLOT_NECK = 2
-        const val SLOT_WEAPON = 3
-        const val SLOT_CHEST = 4
-        const val SLOT_SHIELD = 5
-        const val SLOT_LEG = 7
-        const val SLOT_HAND = 9
-        const val SLOT_FOOT = 10
-        const val SLOT_RING = 12
-        const val SLOT_AMMO = 13
-    }
-}
