+++ game/src/main/kotlin/xlitekt/game/actor/render/block/player/kit/impl/LegInfo.kt
@@ -2,9 +2,9 @@ package xlitekt.game.actor.render.block.player.kit.impl
 
 import io.ktor.utils.io.core.BytePacketBuilder
 import io.ktor.utils.io.core.writeShort
-import xlitekt.game.actor.player.Equipment
 import xlitekt.game.actor.render.Render
 import xlitekt.game.actor.render.block.player.kit.BodyPartInfo
+import xlitekt.game.content.container.equipment.Equipment
 
 class LegInfo : BodyPartInfo(index = 7) {
     override fun equipmentIndex(gender: Render.Appearance.Gender): Int = Equipment.SLOT_LEG
