+++ game/src/main/kotlin/xlitekt/game/actor/render/block/player/kit/impl/WeaponInfo.kt
@@ -1,9 +1,9 @@
 package xlitekt.game.actor.render.block.player.kit.impl
 
 import io.ktor.utils.io.core.BytePacketBuilder
-import xlitekt.game.actor.player.Equipment
 import xlitekt.game.actor.render.Render
 import xlitekt.game.actor.render.block.player.kit.BodyPartInfo
+import xlitekt.game.content.container.equipment.Equipment
 
 class WeaponInfo : BodyPartInfo(index = 3) {
     override fun equipmentIndex(gender: Render.Appearance.Gender): Int = Equipment.SLOT_WEAPON
